Further seperation functionality
